refactor: move lib and gcloudvolume to be available in pipeline
The old ingest folder is defunct.
