+++ python/neuroglancer/pipeline/volumes/gcloudvolume.py
@@ -6,9 +6,9 @@ import numpy as np
 from tqdm import tqdm
 
 import neuroglancer
-import neuroglancer.ingest.lib as lib
-from neuroglancer.ingest.lib import clamp, xyzrange, Vec, Vec3, Bbox, min2, max2, Storage
-from neuroglancer.ingest.volumes import Volume, VolumeCutout, generate_slices
+import neuroglancer.lib as lib
+from neuroglancer.lib import clamp, xyzrange, Vec, Vec3, Bbox, min2, max2, Storage
+from neuroglancer.pipeline.volumes import Volume, VolumeCutout, generate_slices
 from google.cloud import storage as gstorage
 
 class GCloudVolume(Volume):
